requirements.txt
beautifulsoup4==4.12.3
fastapi==0.109.2
feedparser==6.0.11
passlib==1.7.4
pydantic==2.6.1
python-dotenv==1.0.1
python_jose==3.3.0
seleniumbase==4.23.7
SQLAlchemy==2.0.27
uvicorn==0.27.1
